pyproject.toml
[project]
name = "mitre-attack-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=9.1.0",
    "fastapi>=0.116.1",
    "gradio>=5.42.0",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "regex>=2025.7.34",
    "requests>=2.32.4",
    "spacy>=3.8.7",
    "uvicorn>=0.35.0",
]
